% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.variants.guardant.R
\name{gene.variants.guardant}
\alias{gene.variants.guardant}
\title{Abstract Genes with nucleotide variants from Guardant Health genomic reports}
\usage{
gene.variants.guardant(data = dplyr::tibble(Results = clipboard()))
}
\arguments{
\item{data}{The data frame of a genomic report. Ideally this is information copied to the Clipboard from the EHR report by "selecting all" and then "copy". That is the default. If you don't use the clipboard function, you can use a data frame of the text file. The single column data frame should be labeled "Results". Required.}
}
\value{
a data frame with two columns: variables (the redcap variable names) and results (the data to be imported in to redcap)
}
\description{
`gene.variants.guardant()` provides natural language processing tools to abstract genes with nucleotide variants
}
