% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.variants.isolate.oncopanel.R
\name{gene.variants.isolate.oncopanel}
\alias{gene.variants.isolate.oncopanel}
\title{Abstract Genes with Nucleotide variants from Oncopanel genomic reports}
\usage{
gene.variants.isolate.oncopanel(data = dplyr::tibble(Results = clipboard()))
}
\arguments{
\item{data}{The data frame of a genomic report, here an Oncopanel report. Ideally this is information copied to the Clipboard from the EHR report by "selecting all" and then "copy". That is the default. If you don't use the clipboard function, you can use a data frame of the text file. The single column data frame should be labeled "Results". Required.}
}
\value{
a data frame with a column that has the gene names and variants in one cell (i.e. untidy data)
}
\description{
`gene.variants.isolate.oncopanel()` provides natural language processing tools to abstract genes with nucleotide variants from Oncopanel reports
}
\examples{
# Test with embedded data frame "oncopanel_sample_report"
oncopanel_sample_report \%>\%
  gene.variants.isolate.oncopanel()
}
