% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.variants.isolate.snapshot.R
\name{gene.variants.isolate.snapshot}
\alias{gene.variants.isolate.snapshot}
\title{Abstract Genes with genetic variants from SNaPshot genomic reports}
\usage{
gene.variants.isolate.snapshot(data = dplyr::tibble(Results = clipboard()))
}
\arguments{
\item{data}{The data frame of a genomic report. Ideally this is information copied to the Clipboard from the EHR report by "selecting all" and then "copy". That is the default. If you don't use the clipboard function, you can use a data frame of the text file. The single column data frame should be labeled "Results". Required.}
}
\value{
a data frame with a column that has the gene names and variants in one cell
}
\description{
`gene.variants.isolate.snapshot()` provides natural language processing tools to abstract genetic variants and gene names from SNaPshot reports
}
\examples{
# Test with embedded data frame "snapshot_sample_report"
snapshot_sample_report \%>\%
  gene.variants.isolate.snapshot()
}
