% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.variants.R
\name{gene.variants}
\alias{gene.variants}
\title{Abstract Genes with nucleotide variants from genomic reports}
\usage{
gene.variants(data = dplyr::tibble(Results = clipboard()), platform)
}
\arguments{
\item{data}{The data frame of a genomic report. Ideally this is information copied to the Clipboard from the EHR report by "selecting all" and then "copy". That is the default. If you don't use the clipboard function, you can use a data frame of the text file. The single column data frame should be labeled "Results". Required.}
}
\value{
a data frame with two columns: variables (the redcap variable names) and results (the data to be imported in to redcap)
}
\description{
`gene.variants()` integrates various platform-specific NLP functions to text mine gene names and nucleotide variants from genomic reports and transforms them to structured data for import into REDCap
}
