% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.variants.ready.for.redcap.snapshot.R
\name{gene.variants.ready.for.redcap.snapshot}
\alias{gene.variants.ready.for.redcap.snapshot}
\title{Text Mines a SNaPshot genomic report and prepares the data for import into the Genomics Instrument in REDCap}
\usage{
gene.variants.ready.for.redcap.snapshot(
  data = dplyr::tibble(Results = clipboard())
)
}
\arguments{
\item{data}{The data frame of the genomic report of interest, here "SNaPshot". This can be copied to the Clipboard from the EHR report by "selecting all" and then "copy". That is the default. If you don't use the clipboard function, you can use a data frame of the text file. The single column data frame should be labeled "Results". Required..}
}
\value{
a data frame with two columns: "variables" (the redcap variable names) and "results" (the data to be imported in to REDCap). This data can then be combined with the outputs of the other functions of "genetex_to_redcap". That end product would then be imported into REDCap
}
\description{
`gene.variants.ready.for.redcap.snapshot()` text mines a data frame a SNaPshot report so it can be imported into a REDCap "Genomics Instrument" when combined with the other functions of "genetex_to_redcap"
}
\examples{
snapshot_sample_report \%>\%
 gene.variants.ready.for.redcap.snapshot()
}
