% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_regex.R
\name{genes_regex}
\alias{genes_regex}
\title{Create a single regex of all cancer-related gene symbols}
\usage{
genes_regex(file = readr::read_tsv("data-raw/genes.txt"))
}
\arguments{
\item{file}{File of gene symbols. Default imports csv file from data-raw containing over 900 gene symbols. Required}
}
\value{
a single cell string with all gene symbols concatenated by a pipe ("|")
}
\description{
`genes_regex()` produces a regular expression of over 900 HGNC gene names
}
\examples{
#Import "genes" data set from package
genes \%>\%
genes_regex()
}
