% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics.tissue.type.R
\name{genomics.tissue.type}
\alias{genomics.tissue.type}
\title{Assing the appropriate numeric to a string designating the tissue type sequence according to the MCC Patient Registry}
\usage{
genomics.tissue.type(genomics_tissue_type = "")
}
\arguments{
\item{genomics_tissue_type}{A string of the genomics tissue type used in the report (e.g. "primary cutaneous lesion" or "MCCUP" or "met"). Default is blank. Optional.}
}
\value{
A numeric
}
\description{
`genomics.tissue.type()` applies regular expressions to assign a numerical value for the various tissues that may be used for genomic analysis that aligns with the  "genomics_tissue_type" field in the REDCAp Genomics Instrument
}
