% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.gene.variant.variables.R
\name{link.gene.variant.variables}
\alias{link.gene.variant.variables}
\title{Link a column of gene variant variables that are found in the REDCap Genomics Instrument to their appropriate strings}
\usage{
link.gene.variant.variables(data)
}
\arguments{
\item{data}{A Data Frame of Gene Names, Nucleotide Sequences, Amino Acid Sequences and possibly cell-free DNA percentages. This must be a single column data frame, with tidy data (i.e. one name or sequence per cell). These data would like come from a function such as gene.variants.isolate.snapshot()}
}
\value{
a data frame with an additional column with gene variant variables for the REDCap Genomic Instrument. The two columns will be labeled, "genes_variants" and "var". These data could then be used as part of a larger function such as gene.variatns.ready.for.redcap.snapshot()
}
\description{
`link.gene.variant.variables()` creates a column of gene variant variables that are found in the REDCap Genomics Instrument
}
\examples{
# Test with embedded data frame "snapshot_sample_report"
snapshot_sample_report \%>\%
  gene.variants.isolate.snapshot() \%>\%
  link.gene.variant.variables()
}
