% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmr.R
\name{mmr}
\alias{mmr}
\title{Abstract Microsatellite data from genomic reports}
\usage{
mmr(data = dplyr::tibble(Results = readr::clipboard()))
}
\arguments{
\item{data}{The data frame of the genomic report of interest. This can be copied to the Clipboard from the EHR report by "selecting all" and then "copy". That is the default. If you don't use the clipboard function, you can use a data frame of the text file. The single column data frame should be labeled "Results". Required.}
}
\value{
a data frame with two columns: variables (the redcap variable names) and results (the data to be imported in to redcap)
}
\description{
`mmr()` provides NLP tools to text mine mismatch repair status from genomic reports and transforms it to structured data for import into REDCap

`genetex_to_redcap()` provides natural language processing tools to abstract data from a variety of genomic reports and import them to REDCap
}
