% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{platform}
\alias{platform}
\title{Use text mining to assing numeric to genomics platforms}
\usage{
platform(platform)
}
\arguments{
\item{platform}{String of the platform used to generate the genomics report. Acceptable strings include: MGH, SNaPshot, BWH, Oncopanel, Guardant, Guardant Health, Foundation or fmi. Case is not sensitive. Required.}
}
\value{
numeric value corresponding to the various platforms used to generate genomic reports
}
\description{
`platform()` applies regular expressions to assign a numerical value for the various platforms used for genomic reports that aligns with the  "genomics_platform" field in the REDCAp Genomics Instrument
}
