% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile_gene_names.R
\name{reconcile_gene_names}
\alias{reconcile_gene_names}
\title{Replace synonyms for gene names with those that correspond to the data dictionary in the MCC Registry}
\usage{
reconcile_gene_names(data)
}
\arguments{
\item{data}{A data frame from that is the downstream product of the genomics-to-redcap algorithm}
}
\value{
a data frame with gene names replaced
}
\description{
`reconcile_gene_names()` reconciles synonyms import them to REDCap
}
