% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{tmb}
\alias{tmb}
\title{Apply NLP to genomic reports to extract out tumor mutation burden}
\usage{
tmb(data = dplyr::tibble(Results = readr::clipboard()), platform)
}
\arguments{
\item{data}{The data frame of the genomic report of interest. This can be copied to the Clipboard from the EHR report by "selecting all" and then "copy". That is the default. If you don't use the clipboard function, you can use a data frame of the text file. The single column data frame should be labeled "Results". Required.}

\item{platform}{String of the platform used to generate the genomics report. Acceptable strings include: MGH, SNaPshot, BWH, Oncopanel, Guardant, Guardant 360, Foundation or fmi. Case is not sensitive. Required.}
}
\value{
a data frame with two columns: variables (the redcap variable names) and results (the data to be imported in to redcap)
}
\description{
`tmb()` text mines tumor mutation burden (TMB) data from a variety of genomic reports and transforms it to structured data for import into REDCap
}
